build: make generated source files reproducible
This updates the code generation scripts to be deterministic,
which in turn leads to reproducible binaries.

Unfortunately this is incomplete: the exported function
(funcs_metadata.generated.h) and ui event (ui_events_metadata.generated.h)
metadata have some mpack'ed tables, which looks more tricky.

As a temporary workaround, this PR introduces a PRG_GEN_LUA cmake
setting to specify a different lua for those parts of code generation,
so you could plug in a modified build of luajit that preserves table
order. This would be something to fix later.

Co-Authored-By: dundargoc <gocdundar@gmail.com>
Co-Authored-By: Arnout Engelen <arnout@bzzt.net>