build: make generated source files reproducible
This updates the code generation scripts to be deterministic,
which in turn leads to reproducible binaries.

Unfortunately this is incomplete: the exported function
(funcs_metadata.generated.h) and ui event (ui_events_metadata.generated.h)
metadata have some mpack'ed tables, which are currently not
serializaed deterministically due to the fact that lua
table order is (intentionally) not deterministic.

As a temporary workaround, this PR introduces a PRG_GEN_LUA cmake
setting to specify a different lua for those parts of code generation,
so you could plug in a modified build of luajit that preserves table
order. Longer-term we should change the mpack'ed data structure so
it no longer uses tables keyed by strings.

Co-Authored-By: dundargoc <gocdundar@gmail.com>
Co-Authored-By: Arnout Engelen <arnout@bzzt.net>